Fix 32-bit overflow in test
Only a problem when running tests on the pi

+++ internal/deploy/watcher_test.go
@@ -38,9 +38,9 @@ var statusTemplate = `{
 
 type statusData struct {
 	DevArtifact  string
-	DevTime      int
+	DevTime      int64
 	ProdArtifact string
-	ProdTime     int
+	ProdTime     int64
 }
 
 func TestWatch(t *testing.T) {
